requirements.txt
Django==5.2.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0 